requirements.txt
fr24sdk
pandas
pyarrow
airportsdata
requests
boto3
python-dotenv
